:heavy_minus_sign: dep_rm(delete useless dependencies):
