+++ package.json
@@ -49,7 +49,6 @@
     "jest-environment-jsdom-fourteen": "0.1.0",
     "jest-resolve": "24.9.0",
     "jest-watch-typeahead": "0.3.1",
-    "material-table": "^1.50.0",
     "mini-css-extract-plugin": "0.8.0",
     "mui-datatables": "^2.10.2",
     "node-sass": "^4.12.0",
